Changed File/Map Structure for easier scalability
