+++ ResursKollen/View/Employee/EmployeeHomeView.swift
@@ -29,110 +29,6 @@ struct EmployeeHomeView: View {
     }
 }
 
-struct EmployeeMyOrders: View {
-    let user: UserData
-    let orders: [Order]
-
-    var body: some View {
-        ZStack {
-            LinearGradient(
-                gradient: Gradient(colors: [Color(.systemGray6), Color(.systemGray5)]),
-                startPoint: .topLeading,
-                endPoint: .bottomTrailing
-            ).edgesIgnoringSafeArea(.all)
-
-            NavigationView {
-                VStack(alignment: .leading) {
-                    Text("Hej, \(user.name ) 👋")
-                        .font(.largeTitle)
-                        .fontWeight(.bold)
-                        .padding(.top)
-                        .padding(.leading)
-
-
-                    List {
-                        Section(header: Text("Påbörjade ordrar:")      .foregroundColor(.orange)) {
-                            ForEach(orders) { order in
-                                OrderRowMyOrders(order: order)
-                            }
-                        }
-                    }
-                    .listStyle(.insetGrouped)
-                }
-        
-            }
-        }
-    }
-}
-
-struct OrderRowMyOrders: View {
-    let order: Order
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            HStack {
-                VStack(alignment: .leading) {
-                    Text("\(order.description)")
-                        .font(.headline)
-
-                    Text(order.customer.streetName)
-                        .font(.subheadline)
-                        .foregroundColor(.gray)
-                }
-
-                Spacer()
-
-               
-            }
-
-            Text("Förfaller: \(order.dueDate, style: .date)")
-                .font(.caption2)
-                .foregroundColor(.secondary)
-        }
-        .padding(.vertical, 8)
-    }
-}
-
-struct OrderRowAllOrders: View {
-    let order: Order
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            HStack {
-                Image(systemName: "wrench.fill")
-                    .foregroundColor(order.status.color)
-                    .padding(8)
-                    .background(order.status.color.opacity(0.2))
-                    .clipShape(Circle())
-
-                VStack(alignment: .leading) {
-                    Text("\(order.description)")
-                        .font(.headline)
-
-                    Text(order.customer.name)
-                        .font(.subheadline)
-                        .foregroundColor(.gray)
-                }
-
-                Spacer()
-
-            }
-
-            Text("Status: \(order.status.nameSE)")
-                .font(.caption)
-                .padding(.vertical, 4)
-                .padding(.horizontal, 8)
-                .background(order.status.color.opacity(0.1))
-                .foregroundColor(order.status.color)
-                .cornerRadius(8)
-
-            Text("Förfaller: \(order.dueDate, style: .date)")
-                .font(.caption2)
-                .foregroundColor(.secondary)
-        }
-        .padding(.vertical, 8)
-    }
-}
 // Förhandsvisning med mockdata
 #Preview {
     EmployeeHomeView(
